name autogeneration of model and parameterisation adequately rename modules, classes, and adjust the tests
